requirements.txt
openai==1.51.2
langchain==0.3.3
langchain-openai==0.2.2
python-dotenv==1.0.1
langchain_community==0.3.2
faiss-gpu==1.7.2
pypdf==5.0.1